requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM Integration
google-generativeai>=0.3.0

# Environment Management
python-dotenv>=1.0.0

# Report Generation
jinja2>=3.1.0
markdown>=3.5.0

# Utilities
tqdm>=4.65.0
tabulate>=0.9.0
chardet>=5.0.0
